+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 3.4.3 (2025-11-19)
+### Added
+- FakeGato timestamp is coming from Solis data timestamp itself, which is more accurate
+- Added protection for timestamps coming out of order or duplicate ones
+
 ## 3.4.2 (2025-11-18)
 ### Added
-- First stable version
+- First version that will be supported